requirements.txt
Django==4.2.9
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.2.0
django-filter==23.5
openpyxl==3.1.2
reportlab==4.0.8
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
